requirements_joshua.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
argcomplete==1.10.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.8.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
compressed-rtf==1.0.6
comtypes==1.2.0
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
deepl==1.15.0
dill==0.3.6
docx==0.2.4
docx2txt==0.8
ebcdic==1.1.1
entrypoints==0.4
executing==1.2.0
extra-streamlit-components==0.1.56
extract-msg==0.28.7
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.6.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
numpy==1.24.3
olefile==0.46
packaging==23.1
pandas==2.0.1
parso==0.8.3
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
PyAudio==0.2.13
pycryptodome==3.18.0
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.30.0
responses==0.18.0
rich==13.3.5
safetensors==0.3.1
six==1.12.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
stack-data==0.6.2
streamlit==1.22.0
streamlit-authenticator==0.2.2
tenacity==8.2.2
textract==1.6.5
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcag-contrast-ratio==0.9
wcwidth==0.2.6
xlrd==1.2.0
XlsxWriter==3.1.2
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
